Split `main` function code into smaller functions
